Add files via upload
NOTE: none of the work contained in this zip file is my work or my group's own work. It was given to us and we were tasked with getting it working based on the game's rules. The game is Duelyst.
For our work see the Sprint Reports and it-placeholder-preCommit.zip. Our Team name was "Placeholder".
This is the original incomplete code base we were given. 
Note that it is recommended that you try running it with Java version 11 or lower.
You also need to download and install Scala Built Tool (sbt). 
The game runs in the browser.